Refactor ContinueTokenPayload
